requirements.txt
# Core LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# Vector Store & RAG
langchain-chroma>=1.0.0

# Agents
langgraph>=0.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
tqdm>=4.66.0
